eatgo-admin-api/src/main/java/kr/co/fastcampus/eatgo/interfaces/ReviewController.java
package kr.co.fastcampus.eatgo.interfaces;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RestController;

import kr.co.fastcampus.eatgo.application.ReviewService;
import kr.co.fastcampus.eatgo.domain.Review;

@RestController
public class ReviewController {

    @Autowired
    private ReviewService reviewService;

    @GetMapping("/restaurants/{restaurantId}/reviews")
    public List<Review> list(@PathVariable("restaurantId") Long restaurantId) {
        return reviewService.getReviews(restaurantId);
    }
}
